make_band_chimeras

make_band_chimeras Synthesize and store single or multi-band auditory chimeras from pair of original signals.

make_band_chimeras(x1, x2 = "noise", f = 1, n_bands = c(1, 8, 32),
  savepath = ".", bitrate = 16, refilter = FALSE,
  save_waves = TRUE, do_plot = FALSE)

Arguments

x1, x2
original signals. Can be a numeric vector or object of class "Wave"

f
sampling rate in Hz. Not required if using a "Wave" object. Must match for both signals.

n_bands
Number of frequency bands used for synthesis. Can be scalar or vector, in which case one pair of chimeras is synthesized for each element of n_bands.

savepath
path to folder where chimeras will be saved

bitrate
bitrate of signal. 16 used as default, if "Wave" object is used bitrate will be extracted from x1

refilter
set to TRUE to filter again after exchange operation (default FALSE). Not tested.

save_waves
if FALSE waves will be returned in a list

do_plot
plot spectrograms of chimeras. Not implremented.

Source

Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).

Value

wav files of chimera pairs for each band in n_bands

Copyright Bertrand Delgutte, 1999-2000

References

Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a

See also